fix(loop): route short_describe through its deterministic runner on the @task lane (#3570)
short_describe tasks completed with exit 0 but never wrote Ticket.short_description.
The deterministic-phase seam was consulted only by the work-next-headless CLI, while
the lane the scanner's auto-enqueue post_save and the drain safety-net actually use —
core.tasks.execute_headless_task — went straight to the agentic headless runner. That
handed short_describe a generic ticket-work brief its empty toolset cannot satisfy, so
the field stayed blank forever (the scanner's COMPLETED dedup then suppressed re-enqueue).

- Relocate the deterministic-phase seam to teatree.core.deterministic_dispatch (domain):
  a phase->runner registry mirroring headless_dispatch's core->agents inversion, so the
  domain @task worker consults it without a backwards dependency on the interface layer.
- Move the summariser to teatree.agents.short_describe (it calls the one-shot LLM seam,
  which core may not import); agents.apps.ready() registers its runner into the domain seam.
- execute_headless_task now short-circuits through run_deterministic_phase before the
  agentic dispatch — the single seam both headless lanes share, so they never drift.
- ticket_short_describe becomes a thin CLI wrapper; _deterministic_phases is removed.

Regression: tests/teatree_core/test_tasks.py::TestExecuteHeadlessTaskDeterministicPhase —
observed RED (agentic runner called, field blank) before the wiring, GREEN after. RED
re-confirmed this cycle: disabling the short-circuit reproduces get_headless_runner called
+ short_description blank; restoring the fix goes GREEN.

## Architecture pre-check — #3570

## 1. BLUEPRINT § alignment
§17.4 orchestrator-decides / loop-executes + the core->agents runner inversion §6 uses for
headless_dispatch. Claim: a non-agentic Task phase gets a domain-level dispatch seam both
headless lanes share, so neither drifts to the agentic brief its toolset can't satisfy.

## 2. FSM phase boundaries
n/a — no Ticket.State transition added/moved. short_describe is a Task phase, not a lifecycle
state; the Task still runs PENDING -> CLAIMED -> COMPLETED via the existing complete_with_attempt.

## 3. Extension-point contracts
New registry register_deterministic_phase / run_deterministic_phase, mirroring
register_headless_runner. Consumers: teatree.agents.apps.ready() (registers run_short_describe);
core.tasks.execute_headless_task and core/management/commands/tasks.py (work-next-headless CLI)
both consult run_deterministic_phase — the two headless chokepoints.

## 4. Component boundaries
Seam in teatree.core (the domain @task worker consults it); summariser in teatree.agents
(it calls the one-shot LLM seam core may not import). Registration wired at agents app-ready.

## 5. Dependency direction
core -> agents inversion via the registry (no backwards edge; agents registers INTO core at
ready()). `uv run tach check` -> "All modules validated!".

## 6. Test surface
- test_tasks.py::TestExecuteHeadlessTaskDeterministicPhase — @task lane writes short_description
  without agentic dispatch (get_headless_runner not called). Anti-vacuous: RED with the
  short-circuit disabled, GREEN with it.
- test_deterministic_dispatch.py — registry lookup (normalized), None => agentic, raise recorded
  as a durable failed attempt (never left CLAIMED).
- test_short_describe.py — summariser cases (blank-title noop, LLM path, truncation fallback,
  describe_all_missing backfill).

## 7. Resilience invariants
verify-by-re-read: field written via Ticket.objects.update then asserted. idempotency: overwrite
is safe; scanner COMPLETED dedup suppresses re-enqueue once written. fallback-transport:
truncation fallback keeps the field non-blank when the LLM is unavailable. failures recorded
durably via complete_with_attempt (exit 1 + phase_error), never escaping to leave CLAIMED.
sub-agent return contract: runner returns str -> recorded as a structured attempt result dict.

## 8. Identity and key normalization
Phase key canonicalized through normalize_phase at BOTH boundaries (register + lookup) — no
strip-to-match; the normalized phase is the single canonical key.

## 9. Behavior preservation / capability deletion
Replaces the deleted core/management/commands/_deterministic_phases.py (added by #3593). Every
describe case preserved (truncation fallback, blank-title noop, all-missing backfill) with its
tests moved intact; phase_tools keeps short_describe at the empty (_NONE) tool allowance. No
privacy/leak/security gate touched; no must-block test inverted.

## 10. Removability / harness-vs-data
Removable: the seam is a self-contained registry — deleting it reverts to agentic dispatch
(blast radius: 2 call sites + 1 app-ready registration). Harness-vs-data: the phase->runner
mapping lives in data (registry rows populated at app-ready), not hard-coded in the worker —
the worker is the generic mechanism that reads the registry.

Open questions & assumptions:
- The issue posed two candidate shapes (deterministic phase vs. an agentic result-envelope
  field the completion path persists). PR #3593 already chose the deterministic-phase shape
  and wired the CLI lane; this completes it by wiring the @task lane. (status: decided-by-prior-work)
- Assumed the summariser belongs in the agents layer because it calls the one-shot LLM seam,
  which forces the registry inversion (core cannot import agents/interface). (status: assumed)
- Behavior preservation: no privacy/leak/security gate touched; the move preserves every
  describe case (truncation fallback, blank-title noop, all-missing backfill) and its tests,
  and phase_tools keeps short_describe at the empty (_NONE) tool allowance. (status: decided)
- verify-gates FULL push-gate initially failed on the wsgi.py doctest for a missing worktree
  staticfiles/ dir (env artifact — the FULL sweep triggered by the file deletion, not the diff);
  mkdir-ing STATIC_ROOT/staticfiles made verify-gates exit 0. (status: decided — infra, CI provisions collectstatic)
