+++ tests/quality/deferred_import_pegs.toml
@@ -71,7 +71,6 @@
 # god-module split, taking one deferred teatree.core import with them; the four
 # left here (DodE2EOverride, E2EBypassApproval, two get_all_overlays) are real.
 "src/teatree/core/management/commands/ticket.py" = 4
-"src/teatree/core/management/commands/ticket_short_describe.py" = 2
 "src/teatree/core/management/commands/worktree.py" = 1
 "src/teatree/core/managers.py" = 2
 "src/teatree/core/mcp_connectivity.py" = 1
